// Ref/Web/ReferenceDataUpdateService.Web/src/models/RefShippingLineUserView.ts
//------------------------------------------------------------------------------
// <auto-generated>
//     This code was generated from a template.
//
//     Manual changes to this file may cause unexpected behavior in your application.
//     Manual changes to this file will be overwritten if the code is regenerated.
// </auto-generated>
//------------------------------------------------------------------------------
import { IEntity } from "./IEntity";

export class RefShippingLineUserView implements IEntity {
	RSL_PK: string;
	RSL_IsActive: boolean;
	RSL_IsNVO: boolean;
	RSL_CarrierName: string;
	static readonly RSL_CarrierName_MaxLength: number = 75;
	RSL_StandardCarrierAlphaCode: string;
	static readonly RSL_StandardCarrierAlphaCode_MaxLength: number = 4;
	RSL_CargoWiseOneCode: string;
	static readonly RSL_CargoWiseOneCode_MaxLength: number = 4;
	RSL_OceanCarrierMessagingAvailable: boolean;
	RSL_GlobalSailingScheduleAvailable: boolean;
	RSL_ContainerAutomationAvailable: boolean;
	RSL_CargoSphereRatesAvailable: boolean;
	RSL_IsShippingLine: boolean;
	RSL_InvoiceAvailable: boolean;
	RSL_IsCW1User: boolean;
	RSL_EHubIds: string;
	static readonly RSL_EHubIds_MaxLength: number = 250;
	RSL_BookingRequestAvailable: boolean;
	RSL_ShippingInstructionAvailable: boolean;
	RSL_VerifiedGrossContainerWeightAvailable: boolean;
	RSL_ShippingOrderAvailable: boolean;
	RSL_EManifestAvailable: boolean;
	RSL_ShippingLineLogo: Uint8Array;
	RSL_IsSystem: boolean;
	RSL_IsPublished: boolean;
	RSL_IsEditable: boolean;
}